Add csv reader & impl in list cmd
