+++ 01-todo-list/data/csv_reader.go
@@ -1,0 +1,76 @@
+package data
+
+import (
+	"encoding/csv"
+	"log"
+	"os"
+	"strconv"
+
+	"github.com/benfleuty/goprojects/todoapp/model"
+)
+
+type CSVReader struct {
+	FilePath string
+}
+
+func (r *CSVReader) GetAll() []model.Task {
+	records, err := readCsv(r.FilePath)
+	if err != nil {
+		log.Fatalf("There was an error reading the CSV file %s!\n%v", r.FilePath, err)
+	}
+
+	return parseTasks(records)
+}
+
+func parseTasks(records [][]string) []model.Task {
+	if len(records) == 0 {
+		return []model.Task{}
+	}
+
+	var tasks []model.Task
+
+	for _, record := range records {
+		if len(record) < 4 {
+			continue
+		}
+
+		id, err := strconv.Atoi(record[0])
+		if err != nil {
+			continue
+		}
+		description := record[1]
+		created, err := strconv.Atoi(record[2])
+		if err != nil {
+			created = -1
+		}
+		done, err := strconv.ParseBool(record[3])
+		if err != nil {
+			done = false
+		}
+
+		var task model.Task
+		task.ID = id
+		task.Description = description
+		task.Created = created
+		task.Done = done
+		tasks = append(tasks, task)
+	}
+
+	return tasks
+}
+
+func readCsv(path string) ([][]string, error) {
+	f, err := os.Open(path)
+	if err != nil {
+		return nil, err
+	}
+	defer f.Close()
+
+	r := csv.NewReader(f)
+	records, err := r.ReadAll()
+	if err != nil {
+		return nil, err
+	}
+
+	return records, nil
+}
