+++ 01-todo-list/data/task_reader.go
@@ -1,0 +1,7 @@
+package data
+
+import "github.com/benfleuty/goprojects/todoapp/model"
+
+type TaskReader interface {
+	GetAll() []model.Task
+}
